+++ rust/operator-binary/src/operations/graceful_shutdown.rs
@@ -1,5 +1,5 @@
 use snafu::{ResultExt, Snafu};
-use stackable_hdfs_crd::MergedConfig;
+use stackable_hdfs_crd::CommonNodeConfig;
 use stackable_operator::builder::PodBuilder;
 
 #[derive(Debug, Snafu)]
@@ -11,14 +11,14 @@ pub enum Error {
 }
 
 pub fn add_graceful_shutdown_config(
-    merged_config: &(dyn MergedConfig + Send + 'static),
+    merged_config: &CommonNodeConfig,
     pod_builder: &mut PodBuilder,
 ) -> Result<(), Error> {
     // This must be always set by the merge mechanism, as we provide a default value,
     // users can not disable graceful shutdown.
-    if let Some(graceful_shutdown_timeout) = merged_config.graceful_shutdown_timeout() {
+    if let Some(graceful_shutdown_timeout) = merged_config.graceful_shutdown_timeout {
         pod_builder
-            .termination_grace_period(graceful_shutdown_timeout)
+            .termination_grace_period(&graceful_shutdown_timeout)
             .context(SetTerminationGracePeriodSnafu)?;
     }