+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 
 - `operator-rs` `0.56.1` -> `0.57.0` ([#433]).
 - [BREAKING] `.spec.clusterConfig.listenerClass` has been renamed to `.spec.nameNodes.config.listenerClass`, migration will be required when using `external-unstable` ([#450]).
+- Change default value of `dfs.ha.nn.not-become-active-in-safemode` from `true` to `false` ([#458]).
 
 ### Fixed
 
@@ -22,6 +23,7 @@ All notable changes to this project will be documented in this file.
 [#433]: https://github.com/stackabletech/hdfs-operator/pull/433
 [#450]: https://github.com/stackabletech/hdfs-operator/pull/450
 [#451]: https://github.com/stackabletech/hdfs-operator/pull/451
+[#458]: https://github.com/stackabletech/hdfs-operator/pull/458
 
 ## [23.11.0] - 2023-11-24